+++ Backend/SyncService/Models/ArtifactAdded/SizePayload.cs
@@ -2,6 +2,6 @@ namespace SyncService.Models.ArtifactAdded;
 
 public class SizePayload
 {
-    public required int Width { get; set; }
-    public required int Height { get; set; }
+    public required decimal Width { get; set; }
+    public required decimal Height { get; set; }
 }
